Diagram-Aware Image parsing/requirements.txt
Pillow
ultralytics
reportlab
